Mobile - Move From AboutDialog to About Page
