+++ NickvisionApplication.Mobile/NickvisionApplication.Mobile/Views/MainView.axaml.cs
@@ -5,9 +5,11 @@ using Avalonia.Interactivity;
 using Avalonia.Platform.Storage;
 using FluentAvalonia.UI.Controls;
 using NickvisionApplication.Mobile.Controls;
+using NickvisionApplication.Mobile.Helpers;
 using NickvisionApplication.Shared.Controllers;
 using NickvisionApplication.Shared.Events;
 using System;
+using System.Globalization;
 using System.IO;
 using static NickvisionApplication.Shared.Helpers.Gettext;
 
@@ -47,6 +49,23 @@ public partial class MainView : UserControl
         StatusPageHome.Title = _controller.Greeting;
         StatusPageHome.Description = _("Open a folder to get started");
         BtnHomeOpenFolderLabel.Text = _("Open");
+        //About Page
+        LblChangelogHeader.Text = _("Changelog");
+        LblGitHubHeader.Text = _("GitHub Repo");
+        LblReportABug.Text = _("Report a Bug");
+        LblDiscussions.Text = _("Discussions");
+        LblCreditsHeader.Text = _("Credits");
+        LblDescription.Text = _controller.AppInfo.Description;
+        LblVersion.Text = _controller.AppInfo.Version;
+        LblChangelog.Text = _controller.AppInfo.Changelog;
+        var credits = string.Format(_("Developers:\n{0}\n\nDesigners:\n{1}\n\nArtists:\n{2}\n\nTranslators:\n{3}"),
+            string.Join('\n', _controller.AppInfo.Developers.Keys), string.Join('\n', _controller.AppInfo.Designers.Keys),
+            string.Join('\n', _controller.AppInfo.Artists.Keys), _controller.AppInfo.TranslatorCredits);
+        if (string.IsNullOrEmpty(_controller.AppInfo.TranslatorCredits))
+        {
+            credits = credits.Remove(credits.TrimEnd().LastIndexOf('\n'));
+        }
+        LblCredits.Text = credits;
     }
     
     /// <summary>
@@ -149,4 +168,39 @@ public partial class MainView : UserControl
     /// <param name="sender">object?</param>
     /// <param name="e">NavigationViewSelectionChangedEventArgs</param>
     private void NavigationView_OnSelectionChanged(object? sender, NavigationViewSelectionChangedEventArgs e) => ViewStack.CurrentPageName = (e.SelectedItem as NavigationViewItem)!.Tag as string ?? "";
+    
+    /// <summary>
+    /// Occurs when the version button is clicked
+    /// </summary>
+    /// <param name="sender">object?</param>
+    /// <param name="e">RoutedEventArgs</param>
+    private async void CopySystemInformation(object? sender, RoutedEventArgs e)
+    {
+        if (TopLevel.GetTopLevel(this) != null && TopLevel.GetTopLevel(this)!.Clipboard != null)
+        {
+            await TopLevel.GetTopLevel(this)!.Clipboard!.SetTextAsync($"{_controller.AppInfo.ID}\n{_controller.AppInfo.Version}\n\n{Environment.OSVersion}\n{CultureInfo.CurrentCulture.Name}");
+            NotificationSent(sender, new NotificationSentEventArgs(_("Copied system information."), NotificationSeverity.Success));
+        }
+    }
+    
+    /// <summary>
+    /// Occurs when the github repo button is clicked
+    /// </summary>
+    /// <param name="sender">object</param>
+    /// <param name="e">RoutedEventArgs</param>
+    private void GitHubRepo(object sender, RoutedEventArgs e) => _controller.AppInfo.SourceRepo.OpenInBrowser();
+    
+    /// <summary>
+    /// Occurs when the report a bug button is clicked
+    /// </summary>
+    /// <param name="sender">object</param>
+    /// <param name="e">RoutedEventArgs</param>
+    private void ReportABug(object sender, RoutedEventArgs e) => _controller.AppInfo.IssueTracker.OpenInBrowser();
+    
+    /// <summary>
+    /// Occurs when the discussions button is clicked
+    /// </summary>
+    /// <param name="sender">object</param>
+    /// <param name="e">RoutedEventArgs</param>
+    private void Discussions(object sender, RoutedEventArgs e) => _controller.AppInfo.SupportUrl.OpenInBrowser();
 }
