+++ NickvisionApplication.Mobile/NickvisionApplication.Mobile/Views/MainWindow.axaml
@@ -7,6 +7,6 @@
         xmlns:fluentp="using:FluentAvalonia.UI.Controls.Primitives"
         mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="600"
         x:Class="NickvisionApplication.Mobile.Views.MainWindow"
-        MinWidth="320" MinHeight="300" Width="400" Height="600" Icon="../Assets/org.nickvision.application.png">
+        MinWidth="320" MinHeight="300" Width="480" Height="600" Icon="../Assets/org.nickvision.application.png">
     
 </Window>
